Added an adapter class that cleans and transforms the data from github
