+++ src/covidify/sources/data_sources_interface.py
@@ -1,0 +1,10 @@
+import abc
+
+class DataInterface(object, metaclass=abc.ABCMeta):
+
+    def __init__(self):
+        pass
+
+    @abc.abstractmethod
+    def get():
+        raise NotImplementedError('User must define get()')
